Update faq.md with details r.e. spawning
This may be too niche to add, but I spent a fair amount of time and hair pulling trying to understand why spawning a packaged version of my app from a dev version would cause issues.  Maybe a better (or additional) step would be to include docs outlining all the scripts in package.json.